Polish Pass
This commit does a few things (not the ideal practice, but since it's
the last non-documentation commit of the project I'm going to keep it).

This adds an open state property to the lightbox which we toggle along
with the class now. That open state is used largely as a guard for the
keyup listeners this adds for Escape, Left Arrow and Right Arrow keys.
It also smoothes out the next and previous methods so that they update
the image instead of cycling the index values.

Lastly, it installs the Webpack file-loader for using images in the
project. I opted against actually a placeholder image on load, but if I
were to extend this, I'd add icons to the lightbox.